Add msg_filter_ante_handler to block IBCTimeoutOnClose (#1571)
Closes: #XXX

## What is the purpose of the change

Add a msg_filter_ante for v9, and block MsgTimeoutOnClose, due to edge cases that can happen if Terra re-enables the previously closed IBC channels. (This is because Closed channels were never meant to be re-opened)

## Brief Changelog

Adds a ante handler filter for IBC Timeout close messages. This is a port of Bez' old work.

## Testing and Verifying

Covered by provided tests/

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? yes
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? no
  - How is the feature or change documented? TODO