Add osmobuilder image to create binaries for amd64 and arm64 for a release (#1584)
Closes: #1550 

## What is the purpose of the change

This PR proposes a new release process that:

- creates the `osmosisd` binaries for `arm64` and `amd64` statically linked agains `libcosmwasm`
- creates a `releases/` folder with the binaries, archives, source code and sha256checksum

To use it, make sure to be in the root of the repository and run:

```bash
git checkout <TAG>
make -f contrib/images/osmobuilder/Makefile release
```

This will create a `release/` folder with all the appropriate files.

To statically link the binary for different architecture the trick is to build the binary while building the docker image via `docker buildx`.
This process however it's slow when building `arm64` from `amd64` (or viceversa)

## Brief Changelog

- Add `osmobuilder` image to build the images

## Testing and Verifying

To use it, make sure to be in the root of the repository and run:

```bash
git checkout <TAG>
make -f contrib/images/osmobuilder/Makefile release
```

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? no
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? no
  - How is the feature or change documented? not applicable