feat: return bucket index of the current tick from LiquidityPerTickRange query (#6805)
Closes: https://app.clickup.com/t/86a19jj0x

## What is the purpose of the change

This PR is an update to the full range liquidity query. It adds a new return parameter that is the bucket index corresponding to the current tick.

This functionality is necessary for the SQS router in: https://github.com/osmosis-labs/osmosis/pull/6785

By having this index, we don't have to iterate all ticks in search of the current tick. Instead, we can index into the desired bucket in constant time.

## Testing and Verifying

Added unit test and also tested on localosmosis

## Documentation and Release Note

  - [x] Does this pull request introduce a new feature or user-facing behavior changes?
  - [x] Changelog entry added to `Unreleased` section of `CHANGELOG.md`?

Where is the change documented? 
  - [ ] Specification (`x/{module}/README.md`)
  - [ ] Osmosis documentation site
  - [ ] Code comments?
  - [ ] N/A