feat: incentive accumulator migration (#7416)
Closes: #XXX

## What is the purpose of the change

This PR introduces incentive accumulator migration to address the truncation issue with pool 1423.

In the v23 upgrade, we introduce a scaling factor to avoid truncating the per-unit-of-liquidity incentive accumulator. See: https://github.com/osmosis-labs/osmosis/pull/7409

The goal is to have all the newly created pools utilize this accumulator. At the same time, we want to migrate the old pools that are currently affected (e.g. pool 1423). We also want to migrate only the affected pools to limit migration risk.

To achieve all of the above, this PR:
- Introduces a pool ID threshold snapshotted during the upgrade to state. After this threshold, all newly created pools utilize the scaling factor.
- A hard-coded map of pool IDs that are below the threshold but migrated to the scaling factor logic

In the core module logic, we dynamically select the scaling factor based on the pool ID to apply to the accumulator.

Given successful migration, we will migrate the remaining pools in a future upgrade.

### Testnet

Pool ID threshold is selected dynamically based on either mainnet or testnet state.

For any environment other than `osmosis-1`, we avoid performing pool migration to avoid dealing with complexity. This is. fine as we don't care about other environments potentially truncating with older pools.

## Testing and Verifying

- [ ] Test core helpers
- [ ] Test upgrade handler
- [ ] Test genesis

## Documentation and Release Note

  - [ ] Does this pull request introduce a new feature or user-facing behavior changes?
  - [ ] Changelog entry added to `Unreleased` section of `CHANGELOG.md`?

Where is the change documented? 
  - [ ] Specification (`x/{module}/README.md`)
  - [ ] Osmosis documentation site
  - [ ] Code comments?
  - [ ] N/A