refactor: delete unused Power function in osmomath (#9045)
* refactor: comment out unused Power function in osmomath

This commit comments out the unused Power function in osmomath/decimal.go
and its test in decimal_test.go, as this function is not currently in use
and is never called within the codebase. PowerInteger remains intact for
integer exponents.

Closes #3800

* refactor: completely remove Power function documentation

This commit removes the commented out Power function completely, rather
than just commenting it out. This is a cleaner approach as it removes
dead code entirely.